requirements.txt
setuptools
pytest
pyaml
coloredlogs
networkx
validators
jsonschema
flask_restful
pycrypto